refactor(local-community): extract publication-store helpers to local-community/publication-store.ts
Moves storeComment, storeCommentEdit, storeCommentModeration, storeVote,
storeCommunityEditPublication, storePublication, and their supporting
helpers (_prepareCommentWithAnonymity, _prepareCommentEditWithAlias,
_resolveAliasPrivateKeyForCommentPublication, _calculateLinkProps,
_calculateLatestPostProps, _calculateReplyProps, isPublicationReply,
isPublicationPost) into a focused module. Drops 'private' from
_isPublicationAuthorPartOfRoles so the new free functions can reach it
until publication-validation.ts is extracted next.
