+++ src/runtime/node/community/local-community.ts
@@ -239,6 +239,7 @@ import {
     updateInstancePropsWithStartedCommunityOrDb,
     updateInstanceStateWithDbState
 } from "./local-community/db-state.js";
+import { storePublication } from "./local-community/publication-store.js";
 
 // This is a sub we have locally in our pkc datapath, in a NodeJS environment
 export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalCommunityParsedOptions {
@@ -796,481 +797,6 @@ export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalC
         }
     }
 
-    private async storeCommentEdit(
-        commentEditRaw: CommentEditPubsubMessagePublication,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"]
-    ): Promise<undefined> {
-        const log = Logger("pkc-js:local-community:storeCommentEdit");
-        const strippedOutEditPublication = CommentEditPubsubMessagePublicationWithFlexibleAuthorSchema.strip().parse(commentEditRaw); // we strip out here so we don't store any extra props in commentedits table
-        strippedOutEditPublication.author = cleanWireAuthor(strippedOutEditPublication.author); // strip runtime-only author fields (address, publicKey, etc.)
-
-        // Normalize to new wire format: ensure communityPublicKey/communityName for DB columns
-        if (!strippedOutEditPublication.communityPublicKey) strippedOutEditPublication.communityPublicKey = this.signer.address;
-        if (!strippedOutEditPublication.communityName && isStringDomain(this.address))
-            strippedOutEditPublication.communityName = this.address;
-        const commentToBeEdited = this._dbHandler.queryComment(commentEditRaw.commentCid); // We assume commentToBeEdited to be defined because we already tested for its existence above
-        if (!commentToBeEdited) throw Error("The comment to edit doesn't exist"); // unlikely error to happen, but always a good idea to verify
-
-        const editSignedByOriginalAuthor = commentEditRaw.signature.publicKey === commentToBeEdited.signature.publicKey;
-
-        const authorSignerAddress = await getPKCAddressFromPublicKey(commentEditRaw.signature.publicKey);
-
-        const editTableRow = <CommentEditsTableRow>{
-            ...strippedOutEditPublication,
-            isAuthorEdit: editSignedByOriginalAuthor,
-            authorSignerAddress,
-            insertedAt: timestamp()
-        };
-
-        const extraPropsInEdit = remeda
-            .difference(remeda.keys.strict(commentEditRaw), remeda.keys.strict(CommentEditPubsubMessagePublicationSchema.shape))
-            .filter((key) => (key as string) !== "communityAddress"); // communityAddress is excluded because it's been converted to communityPublicKey/communityName above
-        if (extraPropsInEdit.length > 0) {
-            log("Found extra props on CommentEdit", extraPropsInEdit, "Will be adding them to extraProps column");
-            editTableRow.extraProps = remeda.pick(commentEditRaw, extraPropsInEdit);
-        }
-
-        const isEditDuplicate = this._dbHandler.hasCommentEditWithSignatureEncoded(editTableRow.signature.signature);
-        if (isEditDuplicate) {
-            throw new PKCError("ERR_DUPLICATE_COMMENT_EDIT", { editTableRow });
-        }
-
-        this._dbHandler.insertCommentEdits([editTableRow]);
-
-        // If author is deleting a pending or disapproved comment, purge it immediately from the database
-        if (commentEditRaw.deleted === true) {
-            const isPending = commentToBeEdited.pendingApproval;
-            const disapprovalResult = this._dbHandler._queryIsCommentApproved(commentToBeEdited);
-            const isDisapproved = disapprovalResult && !disapprovalResult.approved;
-
-            if (isPending || isDisapproved) {
-                log("Author deleted a pending/disapproved comment, purging immediately", commentEditRaw.commentCid);
-                this._dbHandler.purgeComment(commentEditRaw.commentCid);
-                this._communityUpdateTrigger = true;
-            }
-        }
-    }
-
-    private async storeCommentModeration(
-        commentModRaw: CommentModerationPubsubMessagePublication,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"]
-    ): Promise<undefined> {
-        const log = Logger("pkc-js:local-community:storeCommentModeration");
-        const strippedOutModPublication = CommentModerationPubsubMessagePublicationSchema.strip().parse(commentModRaw); // we strip out here so we don't store any extra props in commentedits table
-        strippedOutModPublication.author = cleanWireAuthor(strippedOutModPublication.author); // strip runtime-only author fields (address, publicKey, etc.)
-
-        // Normalize to new wire format: ensure communityPublicKey/communityName for DB columns
-        if (!strippedOutModPublication.communityPublicKey) strippedOutModPublication.communityPublicKey = this.signer.address;
-        if (!strippedOutModPublication.communityName && isStringDomain(this.address))
-            strippedOutModPublication.communityName = this.address;
-        const commentToBeEdited = this._dbHandler.queryComment(commentModRaw.commentCid); // We assume commentToBeEdited to be defined because we already tested for its existence above
-        if (!commentToBeEdited) throw Error("The comment to edit doesn't exist"); // unlikely error to happen, but always a good idea to verify
-
-        const modSignerAddress = await getPKCAddressFromPublicKey(commentModRaw.signature.publicKey);
-
-        // Determine the target author signer address and domain if this moderation affects the author (ban/flair)
-        let targetAuthorSignerAddress: string | undefined;
-        let targetAuthorDomain: string | undefined;
-        if (strippedOutModPublication.commentModeration.author) {
-            // Check if the comment was published with pseudonymity - if so, get the original author address/domain
-            const aliasInfo = this._dbHandler.queryPseudonymityAliasByCommentCid(commentModRaw.commentCid);
-            if (aliasInfo) {
-                targetAuthorSignerAddress = await getPKCAddressFromPublicKey(aliasInfo.originalAuthorPublicKey);
-                targetAuthorDomain = aliasInfo.originalAuthorName || undefined;
-            } else {
-                targetAuthorSignerAddress = commentToBeEdited.authorSignerAddress;
-                targetAuthorDomain = getAuthorNameFromWire(commentToBeEdited.author);
-            }
-        }
-
-        const modTableRow = <CommentModerationTableRow>{
-            ...strippedOutModPublication,
-            modSignerAddress,
-            insertedAt: timestamp(),
-            targetAuthorSignerAddress,
-            targetAuthorDomain
-        };
-
-        const isCommentModDuplicate = this._dbHandler.hasCommentModerationWithSignatureEncoded(modTableRow.signature.signature);
-        if (isCommentModDuplicate) {
-            throw new PKCError("ERR_DUPLICATE_COMMENT_MODERATION", { modTableRow });
-        }
-
-        const extraPropsInMod = remeda
-            .difference(remeda.keys.strict(commentModRaw), remeda.keys.strict(CommentModerationPubsubMessagePublicationSchema.shape))
-            .filter((key) => (key as string) !== "communityAddress"); // communityAddress is excluded because it's been converted to communityPublicKey/communityName above
-        if (extraPropsInMod.length > 0) {
-            log("Found extra props on CommentModeration", extraPropsInMod, "Will be adding them to extraProps column");
-            modTableRow.extraProps = remeda.pick(commentModRaw, extraPropsInMod);
-        }
-
-        if (modTableRow.commentModeration.purged) {
-            log(
-                "commentModeration.purged=true, and therefore will delete the post/comment and all its reply tree from the db as well as unpin the cids from ipfs",
-                "comment cid is",
-                modTableRow.commentCid
-            );
-
-            const commentToPurge = this._dbHandler.queryComment(modTableRow.commentCid);
-            if (!commentToPurge) throw Error("Comment to purge not found");
-            const purgedTableRows = this._dbHandler.purgeComment(modTableRow.commentCid);
-
-            for (const purgedTableRow of purgedTableRows) await addAllCidsUnderPurgedCommentToBeRemoved(this, purgedTableRow);
-
-            log("Purged comment", modTableRow.commentCid, "and its comment and comment update children", "out of DB and IPFS");
-
-            await rmUnneededMfsPaths(this); // not sure if needed here
-            if (this.updateCid) {
-                // need to remove any update cids with reference to purged comment
-                this._blocksToRm.push(this.updateCid);
-                this._cidsToUnPin.add(this.updateCid);
-            }
-        } else if ("approved" in modTableRow.commentModeration) {
-            if (modTableRow.commentModeration.approved) {
-                log(
-                    "commentModeration.approved=true, and therefore move comment from pending approval and add it to IPFS",
-                    "comment cid is",
-                    modTableRow.commentCid
-                );
-
-                await this._addCommentRowToIPFS(
-                    commentToBeEdited,
-                    Logger("pkc-js:local-community:storeCommentModeration:_addCommentRowToIPFS")
-                );
-                this._dbHandler.approvePendingComment({ cid: modTableRow.commentCid });
-            } else {
-                const shouldPurgeDisapprovedComment = Object.keys(modTableRow.commentModeration).length === 1; // no other props were included, if so purge the comment
-                log(
-                    "commentModeration.approved=false, and therefore this comment will be removed entirely from DB",
-                    "should we purge this comment? = ",
-                    shouldPurgeDisapprovedComment,
-                    "comment cid is",
-                    modTableRow.commentCid
-                );
-                if (shouldPurgeDisapprovedComment) this._dbHandler.purgeComment(modTableRow.commentCid);
-                else this._dbHandler.removeCommentFromPendingApproval({ cid: modTableRow.commentCid });
-            }
-        }
-        this._dbHandler.insertCommentModerations([modTableRow]);
-        this._communityUpdateTrigger = true;
-        log("Inserted comment moderation", "of comment", modTableRow.commentCid, "into db", "with props", modTableRow);
-    }
-
-    private async storeVote(
-        newVoteProps: VotePubsubMessagePublication,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"]
-    ) {
-        const log = Logger("pkc-js:local-community:storeVote");
-
-        const authorSignerAddress = await getPKCAddressFromPublicKey(newVoteProps.signature.publicKey);
-        this._dbHandler.deleteVote(authorSignerAddress, newVoteProps.commentCid);
-        const voteTableRow = <VotesTableRow>{
-            ...remeda.pick(newVoteProps, ["vote", "commentCid", "protocolVersion", "timestamp"]),
-            authorSignerAddress,
-            insertedAt: timestamp()
-        };
-        const extraPropsInVote = remeda.difference(
-            remeda.keys.strict(newVoteProps),
-            remeda.keys.strict(VotePubsubMessagePublicationSchema.shape)
-        );
-        if (extraPropsInVote.length > 0) {
-            log("Found extra props on Vote", extraPropsInVote, "Will be adding them to extraProps column");
-            voteTableRow.extraProps = remeda.pick(newVoteProps, extraPropsInVote);
-        }
-
-        this._dbHandler.insertVotes([voteTableRow]);
-        log("Inserted vote", "of comment", voteTableRow.commentCid, "into db", "with props", voteTableRow);
-        return undefined;
-    }
-
-    private async storeCommunityEditPublication(
-        editProps: CommunityEditPubsubMessagePublication,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"]
-    ) {
-        const log = Logger("pkc-js:local-community:storeCommunityEdit");
-
-        const authorSignerAddress = await getPKCAddressFromPublicKey(editProps.signature.publicKey);
-        const authorIdentity = getAuthorNameFromWire(editProps.author) || authorSignerAddress;
-        log(
-            "Received community edit",
-            editProps.communityEdit,
-            "from author",
-            authorIdentity,
-            "with signer address",
-            authorSignerAddress,
-            "Will be using these props to edit the community props"
-        );
-
-        const propsAfterEdit = remeda.pick(this, remeda.keys.strict(editProps.communityEdit));
-        log("Current props from community edit (not edited yet)", propsAfterEdit);
-        lodashDeepMerge(propsAfterEdit, editProps.communityEdit);
-        await this.edit(propsAfterEdit);
-        return undefined;
-    }
-
-    private isPublicationReply(publication: CommentPubsubMessagePublication): publication is ReplyPubsubMessageWithCommunityAuthor {
-        return Boolean(publication.parentCid);
-    }
-
-    private isPublicationPost(publication: CommentPubsubMessagePublication): publication is PostPubsubMessageWithCommunityAuthor {
-        return !publication.parentCid;
-    }
-
-    private async _calculateLinkProps(
-        link: CommentPubsubMessagePublication["link"]
-    ): Promise<Pick<CommentIpfsType, "thumbnailUrl" | "thumbnailUrlWidth" | "thumbnailUrlHeight"> | undefined> {
-        if (!link || !this.settings?.fetchThumbnailUrls) return undefined;
-        return getThumbnailPropsOfLink(link, this, this.settings.fetchThumbnailUrlsProxyUrl);
-    }
-
-    private async _calculateLatestPostProps(): Promise<Pick<CommentIpfsType, "previousCid" | "depth">> {
-        this._dbHandler.createTransaction();
-        const previousCid = this._dbHandler.queryLatestPostCid()?.cid;
-        this._dbHandler.commitTransaction();
-        return { depth: 0, previousCid };
-    }
-
-    private async _calculateReplyProps(
-        comment: CommentPubsubMessagePublication
-    ): Promise<Pick<CommentIpfsType, "previousCid" | "depth" | "postCid">> {
-        if (!comment.parentCid) throw Error("Reply has to have parentCid");
-
-        this._dbHandler.createTransaction();
-        const commentsUnderParent = this._dbHandler.queryCommentsUnderComment(comment.parentCid);
-        const parent = this._dbHandler.queryComment(comment.parentCid);
-        this._dbHandler.commitTransaction();
-
-        if (!parent) throw Error("Failed to find parent of reply");
-
-        return {
-            depth: parent.depth + 1,
-            postCid: parent.postCid,
-            previousCid: commentsUnderParent[0]?.cid
-        };
-    }
-
-    private async _resolveAliasPrivateKeyForCommentPublication(opts: {
-        mode: PseudonymityAliasRow["mode"];
-        originalAuthorPublicKey: PseudonymityAliasRow["originalAuthorPublicKey"];
-        postCid?: string;
-    }): Promise<string> {
-        if (opts.mode === "per-post") {
-            // For a new post (no postCid yet), always generate a fresh alias; once stored the postCid will be used for reuse.
-            if (opts.postCid) {
-                const existing = this._dbHandler.queryPseudonymityAliasForPost(opts.originalAuthorPublicKey, opts.postCid);
-                if (existing?.aliasPrivateKey) return existing.aliasPrivateKey;
-            }
-            return (await this._pkc.createSigner()).privateKey;
-        } else if (opts.mode === "per-reply") {
-            const signer = await this._pkc.createSigner();
-            return signer.privateKey;
-        } else if (opts.mode === "per-author") {
-            const existing = this._dbHandler.queryPseudonymityAliasForAuthor(opts.originalAuthorPublicKey);
-            if (existing?.aliasPrivateKey) return existing.aliasPrivateKey;
-            const signer = await this._pkc.createSigner();
-            return signer.privateKey;
-        } else throw Error(`Unsupported pseudonymityMode (${opts.mode})`);
-    }
-
-    private async _prepareCommentWithAnonymity(originalComment: CommentPubsubMessagePublication): Promise<{
-        publication: CommentPubsubMessagePublication;
-        anonymity?: {
-            aliasPrivateKey: PseudonymityAliasRow["aliasPrivateKey"];
-            originalAuthorPublicKey: PseudonymityAliasRow["originalAuthorPublicKey"];
-            mode: PseudonymityAliasRow["mode"];
-            originalComment: CommentPubsubMessagePublication;
-        };
-    }> {
-        const mode = this.features?.pseudonymityMode;
-        if (!mode) return { publication: originalComment };
-
-        // Mods (owner, admin, moderator) are never pseudonymized
-        const isAuthorMod = await this._isPublicationAuthorPartOfRoles(originalComment, ["owner", "admin", "moderator"]);
-        if (isAuthorMod) return { publication: originalComment };
-
-        const originalAuthorPublicKey = originalComment.signature.publicKey;
-        const postCid = originalComment.postCid;
-        const aliasPrivateKey = await this._resolveAliasPrivateKeyForCommentPublication({
-            mode,
-            originalAuthorPublicKey,
-            postCid
-        });
-        const aliasSigner = await this._pkc.createSigner({ privateKey: aliasPrivateKey, type: "ed25519" });
-        const displayName = originalComment.author?.displayName;
-        const sanitizedAuthor = cleanWireAuthor(displayName !== undefined ? { displayName } : undefined);
-
-        const anonymizedComment = remeda.clone(originalComment);
-
-        if (sanitizedAuthor !== undefined) {
-            anonymizedComment.author = sanitizedAuthor;
-        } else {
-            delete anonymizedComment.author;
-        }
-        anonymizedComment.signature = await signComment({
-            comment: { ...anonymizedComment, signer: aliasSigner, communityAddress: this.address },
-            pkc: this._pkc
-        });
-
-        return {
-            publication: anonymizedComment,
-            anonymity: {
-                aliasPrivateKey,
-                originalAuthorPublicKey,
-                mode,
-                originalComment
-            }
-        };
-    }
-
-    private async _prepareCommentEditWithAlias(originalEdit: CommentEditPubsubMessagePublication) {
-        const aliasSignerOfComment = this._dbHandler.queryPseudonymityAliasByCommentCid(originalEdit.commentCid);
-        if (!aliasSignerOfComment) return originalEdit;
-
-        const aliasSigner = await this._pkc.createSigner({
-            privateKey: aliasSignerOfComment.aliasPrivateKey,
-            type: "ed25519"
-        });
-        const commentEditSignedByAlias = remeda.clone(originalEdit);
-        delete commentEditSignedByAlias.author;
-        commentEditSignedByAlias.signature = await signCommentEdit({
-            edit: { ...commentEditSignedByAlias, signer: aliasSigner, communityAddress: this.address },
-            pkc: this._pkc
-        });
-
-        return commentEditSignedByAlias;
-    }
-
-    private async storeComment(opts: {
-        commentPubsub: CommentPubsubMessagePublication;
-        pendingApproval?: boolean;
-        pseudonymityMode?: PseudonymityAliasRow["mode"];
-        originalCommentSignatureEncoded?: string;
-    }): Promise<{ comment: CommentIpfsType; cid: CommentUpdateType["cid"] }> {
-        const { commentPubsub, pendingApproval, pseudonymityMode, originalCommentSignatureEncoded } = opts;
-        const log = Logger("pkc-js:local-community:handleChallengeExchange:storeComment");
-
-        const commentIpfs = <CommentIpfsType>{
-            ...commentPubsub,
-            ...(await this._calculateLinkProps(commentPubsub.link)),
-            ...(this.isPublicationPost(commentPubsub) && (await this._calculateLatestPostProps())),
-            ...(this.isPublicationReply(commentPubsub) && (await this._calculateReplyProps(commentPubsub))),
-            ...(pseudonymityMode ? { pseudonymityMode } : {})
-        };
-
-        // Normalize to new wire format: ensure communityPublicKey/communityName, remove old communityAddress
-        commentIpfs.communityPublicKey = this.signer.address;
-        if (isStringDomain(this.address)) commentIpfs.communityName = this.address;
-        delete (commentIpfs as Record<string, unknown>).communityAddress;
-
-        // Strip runtime-only author fields (nameResolved, address, publicKey, etc.) before IPFS storage
-        commentIpfs.author = cleanWireAuthor(commentIpfs.author);
-
-        const ipfsClient = this._clientsManager.getDefaultKuboRpcClient();
-
-        const file = pendingApproval
-            ? undefined
-            : await retryKuboIpfsAddAndProvide({
-                  ipfsClient: ipfsClient._client,
-                  log,
-                  content: deterministicStringify(commentIpfs),
-                  addOptions: { pin: true },
-                  provideOptions: { recursive: true },
-                  provideInBackground: false
-              });
-
-        const commentCid = file?.path || (await calculateIpfsCidV0(deterministicStringify(commentIpfs)));
-        const postCid = commentIpfs.postCid || commentCid; // if postCid is not defined, then we're adding a post to IPFS, so its own cid is the postCid
-        const authorSignerAddress = await getPKCAddressFromPublicKey(commentPubsub.signature.publicKey);
-
-        const strippedOutCommentIpfs = CommentIpfsSchema.strip().parse(commentIpfs); // remove unknown props
-        strippedOutCommentIpfs.author = cleanWireAuthor(strippedOutCommentIpfs.author); // strip runtime-only author fields (address, publicKey, etc.)
-
-        const signaturesToCheck = Array.from(
-            new Set(
-                [commentPubsub.signature.signature, originalCommentSignatureEncoded].filter((sig): sig is string => typeof sig === "string")
-            )
-        );
-        const isCommentDuplicate = signaturesToCheck.some((signatureEncoded) =>
-            this._dbHandler.hasCommentWithSignatureEncoded(signatureEncoded)
-        );
-        if (isCommentDuplicate) {
-            this._cidsToUnPin.add(commentCid);
-            throw new PKCError("ERR_DUPLICATE_COMMENT", { file, commentIpfs, commentPubsub });
-        }
-
-        const commentRow = <CommentsTableRow>{
-            ...strippedOutCommentIpfs,
-            cid: commentCid,
-            postCid,
-            authorSignerAddress,
-            insertedAt: timestamp(),
-            pendingApproval
-        };
-
-        const unknownProps = remeda
-            .difference(remeda.keys.strict(commentPubsub), remeda.keys.strict(CommentPubsubMessagePublicationSchema.shape))
-            .filter((key) => (key as string) !== "communityAddress"); // communityAddress is excluded because it's been converted to communityPublicKey/communityName above
-
-        if (unknownProps.length > 0) {
-            log("Found extra props on Comment", unknownProps, "Will be adding them to extraProps column");
-            commentRow.extraProps = remeda.pick(commentPubsub, unknownProps);
-        }
-        if (originalCommentSignatureEncoded) commentRow.originalCommentSignatureEncoded = originalCommentSignatureEncoded;
-
-        // we may need to query comment and verify its signature
-        this._dbHandler.createTransaction();
-        try {
-            if (!pendingApproval) {
-                const { number, postNumber } = this._dbHandler.getNextCommentNumbers(commentRow.depth);
-                commentRow.number = number;
-                if (typeof postNumber === "number") commentRow.postNumber = postNumber;
-            }
-            this._dbHandler.insertComments([commentRow]);
-            if (typeof this.settings?.maxPendingApprovalCount === "number")
-                this._dbHandler.removeOldestPendingCommentIfWeHitMaxPendingCount(this.settings.maxPendingApprovalCount);
-            this._dbHandler.commitTransaction();
-        } catch (e) {
-            this._dbHandler.rollbackTransaction();
-            throw e;
-        }
-        log("Inserted comment", commentRow.cid, "into db", "with props", commentRow);
-
-        return { comment: commentIpfs, cid: commentCid };
-    }
-
-    private async storePublication(request: DecryptedChallengeRequestMessageType, pendingApproval?: boolean) {
-        if (request.vote) return this.storeVote(request.vote, request.challengeRequestId);
-        else if (request.commentEdit) {
-            const commentEditWithAlias = await this._prepareCommentEditWithAlias(request.commentEdit);
-            return this.storeCommentEdit(commentEditWithAlias, request.challengeRequestId);
-        } else if (request.commentModeration) return this.storeCommentModeration(request.commentModeration, request.challengeRequestId);
-        else if (request.comment) {
-            const originalCommentSignatureEncoded = request.comment.signature.signature;
-            const { publication, anonymity } = await this._prepareCommentWithAnonymity(request.comment);
-            const storedComment = await this.storeComment({
-                commentPubsub: publication,
-                pendingApproval,
-                pseudonymityMode: anonymity?.mode,
-                originalCommentSignatureEncoded: anonymity ? originalCommentSignatureEncoded : undefined
-            });
-
-            if (anonymity)
-                this._dbHandler.insertPseudonymityAliases([
-                    {
-                        commentCid: storedComment.cid,
-                        aliasPrivateKey: anonymity.aliasPrivateKey,
-                        originalAuthorPublicKey: anonymity.originalAuthorPublicKey,
-                        originalAuthorName: getAuthorNameFromWire(anonymity.originalComment.author) || null,
-                        mode: anonymity.mode,
-                        insertedAt: timestamp()
-                    }
-                ]);
-
-            return storedComment;
-        } else if (request.communityEdit) return this.storeCommunityEditPublication(request.communityEdit, request.challengeRequestId);
-        else throw Error("Don't know how to store this publication" + request);
-    }
-
     private async _decryptOrRespondWithFailure(request: ChallengeRequestMessageType | ChallengeAnswerMessageType): Promise<string> {
         const log = Logger("pkc-js:local-community:_decryptOrRespondWithFailure");
         try {
@@ -1492,7 +1018,7 @@ export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalC
         request: DecryptedChallengeRequestMessageType,
         pendingApproval?: boolean
     ): Promise<(DecryptedChallengeVerification & Required<Pick<DecryptedChallengeVerificationMessageType, "encrypted">>) | undefined> {
-        const commentAfterAddingToIpfs = await this.storePublication(request, pendingApproval);
+        const commentAfterAddingToIpfs = await storePublication(this, request, pendingApproval);
         if (!commentAfterAddingToIpfs) return undefined;
         const authorSignerAddress = await getPKCAddressFromPublicKey(commentAfterAddingToIpfs.comment.signature.publicKey);
         const authorDomain = getAuthorNameFromWire(commentAfterAddingToIpfs.comment.author);
@@ -1587,7 +1113,7 @@ export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalC
         }
     }
 
-    private async _isPublicationAuthorPartOfRoles(
+    async _isPublicationAuthorPartOfRoles(
         publication: Pick<CommentModerationPubsubMessagePublication, "author" | "signature">,
         rolesToCheckAgainst: CommunityRoleNameUnion[]
     ): Promise<boolean> {
